+++ src/server/common.ts
@@ -2,6 +2,7 @@ import { BadRequestError } from "../common/errors";
 import { checkSourcifyChainId } from "../sourcify-chains";
 import {
   CheckedContract,
+  Match,
   PathContent,
   Status,
 } from "@ethereum-sourcify/lib-sourcify";
@@ -96,4 +97,34 @@ declare module "express-session" {
     contractWrappers: ContractWrapperMap;
     unusedSources: string[];
   }
+}
+
+export interface ResponseMatch
+  extends Omit<Match, "runtimeMatch" | "creationMatch"> {
+  status: Status;
+}
+
+export function getMatchStatus(match: Match): Status {
+  if (match.runtimeMatch === "perfect" || match.creationMatch === "perfect") {
+    return "perfect";
+  }
+  if (match.runtimeMatch === "partial" || match.creationMatch === "partial") {
+    return "partial";
+  }
+  if (match.runtimeMatch === "extra-file-input-bug") {
+    return "extra-file-input-bug";
+  }
+  return null;
+}
+
+export function getResponseMatchFromMatch(match: Match): ResponseMatch {
+  const status = getMatchStatus(match);
+  const responseMatch = {
+    ...match,
+    status,
+    runtimeMatch: undefined,
+    creationMatch: undefined,
+  };
+
+  return responseMatch;
 }