+++ src/server/controllers/verification/create2/stateless/create2.stateless.handlers.ts
@@ -11,6 +11,7 @@ import {
 } from "@ethereum-sourcify/lib-sourcify";
 import { BadRequestError } from "../../../../../common/errors";
 import { Create2VerifyRequest } from "../create2.common";
+import { getResponseMatchFromMatch } from "../../../../common";
 
 export async function verifyCreate2Handler(
   req: Create2VerifyRequest,
@@ -56,9 +57,9 @@ export async function verifyCreate2Handler(
     abiEncodedConstructorArguments
   );
 
-  if (match.status) {
+  if (match.runtimeMatch || match.creationMatch) {
     await services.repository.storeMatch(contract, match);
   }
 
-  res.send({ result: [match] });
+  res.send({ result: [getResponseMatchFromMatch(match)] });
 }