+++ src/server/services/VerificationService.ts
@@ -5,9 +5,9 @@ import {
   Match,
   /* ContextVariables, */
 } from "@ethereum-sourcify/lib-sourcify";
-import { SourcifyEventManager } from "../../common/SourcifyEventManager/SourcifyEventManager";
-import { getCreatorTx } from "./VerificationService-util";
-import { supportedChainsMap } from "../../sourcify-chains";
+import { getCreatorTx } from "./utils/contract-creation-util";
+import { ContractIsAlreadyBeingVerifiedError } from "../../common/errors/ContractIsAlreadyBeingVerifiedError";
+import { logger } from "../../common/loggerLoki";
 
 export interface IVerificationService {
   supportedChainsMap: SourcifyChainMap;
@@ -22,9 +22,27 @@ export interface IVerificationService {
 
 export class VerificationService implements IVerificationService {
   supportedChainsMap: SourcifyChainMap;
+  activeVerificationsByChainIdAddress: {
+    [chainIdAndAddress: string]: boolean;
+  } = {};
 
   constructor(supportedChainsMap: SourcifyChainMap) {
     this.supportedChainsMap = supportedChainsMap;
+  }
+
+  private throwIfContractIsAlreadyBeingVerified(
+    chainId: string,
+    address: string
+  ) {
+    if (
+      this.activeVerificationsByChainIdAddress[`${chainId}:${address}`] !==
+      undefined
+    ) {
+      logger.warn(
+        `The contract ${address} on chainId ${chainId} is already being verified, please wait`
+      );
+      throw new ContractIsAlreadyBeingVerifiedError(chainId, address);
+    }
   }
 
   public async verifyDeployed(
@@ -34,37 +52,29 @@ export class VerificationService implements IVerificationService {
     /* contextVariables?: ContextVariables, */
     creatorTxHash?: string
   ): Promise<Match> {
+    this.throwIfContractIsAlreadyBeingVerified(chainId, address);
+    this.activeVerificationsByChainIdAddress[`${chainId}:${address}`] = true;
+
     const sourcifyChain = this.supportedChainsMap[chainId];
-    let match;
+    const foundCreatorTxHash =
+      creatorTxHash ||
+      (await getCreatorTx(sourcifyChain, address)) ||
+      undefined;
+
+    /* eslint-disable no-useless-catch */
     try {
-      match = await verifyDeployed(
+      return await verifyDeployed(
         checkedContract,
         sourcifyChain,
         address,
         /* contextVariables, */
-        creatorTxHash
+        foundCreatorTxHash
       );
-      return match;
-    } catch (err) {
-      // Find the creator tx if it wasn't supplied and try verifying again with it.
-      if (
-        !creatorTxHash &&
-        err instanceof Error &&
-        err.message === "The deployed and recompiled bytecode don't match."
-      ) {
-        const foundCreatorTxHash = await getCreatorTx(sourcifyChain, address);
-        if (foundCreatorTxHash) {
-          match = await verifyDeployed(
-            checkedContract,
-            sourcifyChain,
-            address,
-            /* contextVariables, */
-            foundCreatorTxHash
-          );
-          return match;
-        }
-      }
-      throw err;
+    } catch (e) {
+      throw e;
+    } finally {
+      delete this.activeVerificationsByChainIdAddress[`${chainId}:${address}`];
     }
   }
+  /* eslint-enable no-useless-catch */
 }